#7 add: hooks
