+++ src/components/MailForm/MailForm.tsx
@@ -12,27 +12,11 @@ import {
 } from "../ui/form";
 import { Input } from "../ui/input";
 import { Button } from "../ui/button";
-import { useForm } from "react-hook-form";
-import { zodResolver } from "@hookform/resolvers/zod";
-import { formSchema } from "@/lib/formSchema";
 import { Textarea } from "../ui/textarea";
+import { useMailForm } from "@/hooks/useMailForm";
 
 const MailForm = () => {
-  const form = useForm({
-    resolver: zodResolver(formSchema),
-    defaultValues: {
-      username: "",
-      subject: "",
-      email: "",
-      content: "",
-    },
-  });
-
-  function onSubmit(values: any) {
-    // Do something with the form values.
-    // ✅ This will be type-safe and validated.
-    console.log(values);
-  }
+  const { form, onSubmit } = useMailForm();
 
   return (
     <div>
